% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{add_dataframe}
\alias{add_dataframe}
\title{Add data frame to csvw table annotation}
\usage{
add_dataframe(table, filename, default_schema)
}
\arguments{
\item{table}{a `csvw:Table` annotation}

\item{filename}{a filename/ URL for the csv table}

\item{default_schema}{an optional fallback used if the table specification
doesn't include a tableSchema}
}
\value{
a table annotation with a `dataframe` attribute added with data frame
holding the contents of the table
}
\description{
Add data frame to csvw table annotation
}
