% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{base_uri}
\alias{base_uri}
\title{Retrieve the base URI from configuration}
\usage{
base_uri()
}
\value{
returns the value of \code{csvwr_base_uri} option, defaulting to \code{example.net}
}
\description{
Retrieve the base URI from configuration
}
\examples{
\dontrun{
base_uri() # returns default

options(csvwr_base_uri="http://www.w3.org/2013/csvw/tests/")
base_uri()
}
}
