% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{base_url}
\alias{base_url}
\title{Determine the base URL for CSVW metadata}
\usage{
base_url(metadata, location)
}
\arguments{
\item{metadata}{the csvw metadata}

\item{location}{where the metadata was originally located}
}
\value{
A string containing the base URL
}
\description{
Determine the base URL for CSVW metadata
}
