% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{coalesce_truth}
\alias{coalesce_truth}
\title{Coalesce value to truthiness}
\usage{
coalesce_truth(x)
}
\arguments{
\item{x}{logical, `NA` or `NULL`}
}
\value{
`FALSE` if x is anything but `TRUE`
}
\description{
Determineness whether the input is true, with missing values being interpreted as false.
}
