% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{csvwr_example}
\alias{csvwr_example}
\title{Get path to csvwr example}
\usage{
csvwr_example(path = NULL)
}
\arguments{
\item{path}{The filename. If \code{NULL}, the example files will be listed.}
}
\value{
either a file path or a vector of filenames
}
\description{
The csvwr package includes some example csvw files in it's \code{inst/extdata} directory.
You can use this function to find them.
}
\details{
Inspired by \code{\link[readr:readr_example]{readr::readr_example()}}
}
\examples{
csvwr_example()
csvwr_example("computer-scientists.csv")
}
