% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\docType{data}
\name{default_dialect}
\alias{default_dialect}
\title{CSVW default dialect}
\format{
An object of class \code{list} of length 13.
}
\usage{
default_dialect
}
\value{
a list specifying a default csv dialect
}
\description{
The [CSVW Default Dialect specification](https://w3c.github.io/csvw/metadata/#dialect-descriptions)
described in [CSV Dialect Description Format](http://dataprotocols.org/csv-dialect/).
}
\keyword{datasets}
