% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{derive_metadata}
\alias{derive_metadata}
\title{Derive csvw metadata from a csv file}
\usage{
derive_metadata(filename)
}
\arguments{
\item{filename}{a csv file}
}
\value{
a list of csvw metadata
}
\description{
Derive csvw metadata from a csv file
}
\examples{
derive_metadata(csvwr_example("computer-scientists.csv"))
}
