% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{derive_table_schema}
\alias{derive_table_schema}
\title{Derive csvw table schema from a data frame}
\usage{
derive_table_schema(d)
}
\arguments{
\item{d}{a data frame}
}
\value{
a list describing a \code{csvw:tableSchema}
}
\description{
Derive csvw table schema from a data frame
}
\examples{
derive_table_schema(data.frame(a=1,b=2))
}
