% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{extract_table}
\alias{extract_table}
\title{Extract a referenced table from CSVW metadata}
\usage{
extract_table(csvw, reference)
}
\arguments{
\item{csvw}{the metadata}

\item{reference}{a foreign key reference expressed as a list containing either a
reference attribute or a schemaReference attribute}
}
\value{
a csvw table
}
\description{
Extract a referenced table from CSVW metadata
}
