% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{find_existing_file}
\alias{find_existing_file}
\title{Find the first existing file from a set of candidates}
\usage{
find_existing_file(filenames)
}
\arguments{
\item{filenames}{a vector of candidates}
}
\value{
If one of the filenames passed is found, then the first is returned.
If none of the filenames exist, \code{NULL} is returned
}
\description{
Find the first existing file from a set of candidates
}
