% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{find_metadata}
\alias{find_metadata}
\title{Find metadata for a tabular file}
\usage{
find_metadata(filename)
}
\arguments{
\item{filename}{a csv file}
}
\value{
a uri for the metadata, or null if none were found
}
\description{
Searches through the default locations attempting to locate metadata.
}
\keyword{internal}
