% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{is_absolute_url}
\alias{is_absolute_url}
\title{Does the string provide an absolute URL}
\usage{
is_absolute_url(string)
}
\arguments{
\item{string}{the url, path or template}
}
\value{
true if the string is an absolute url
}
\description{
Does the string provide an absolute URL
}
\keyword{internal}
