% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_property}
\alias{is_property}
\title{Identify csvw properties}
\usage{
is_property(x)
}
\arguments{
\item{x}{a vector}
}
\value{
A boolean
}
\description{
Overrides the default behaviour in \code{purrr::pluck_depth} so that we don't descend into errors
}
\keyword{internal}
