% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{list_of_lists_to_df}
\alias{list_of_lists_to_df}
\title{Parse list of lists specification into a data frame}
\usage{
list_of_lists_to_df(ll)
}
\arguments{
\item{ll}{a list of lists}
}
\value{
a data frame with a row per list
}
\description{
Parse list of lists specification into a data frame
}
