% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{locate_metadata}
\alias{locate_metadata}
\title{Locate metadata for a table}
\usage{
locate_metadata(filename, metadata)
}
\arguments{
\item{filename}{a path for a csv table or a json metadata document}

\item{metadata}{optional user metadata}
}
\value{
csvw metadata list
}
\description{
Follows the procedure defined in the \href{https://w3c.github.io/csvw/syntax/#locating-metadata}{csvw model}:
}
\details{
\enumerate{
\item Metadata supplied by the user
\item Metadata referenced by a link header
\item Metadata located through default paths
\item Metadata embedded in the file
}

We extend this to use the \link{derive_metadata} function to inspect the table itself.
}
\keyword{internal}
