% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{location_configuration}
\alias{location_configuration}
\title{Identify metadata location configurations for a tabular file}
\usage{
location_configuration(filename)
}
\arguments{
\item{filename}{a csv file}
}
\value{
a character vector of URI templates
}
\description{
Returns default locations. Will ultimately retrieve remote configuration
}
