% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{normalise_metadata}
\alias{normalise_metadata}
\title{Normalise metadata}
\usage{
normalise_metadata(metadata, location)
}
\arguments{
\item{metadata}{a csvw metadata list}

\item{location}{the location of the metadata}
}
\value{
metadata coerced into a
\href{https://www.w3.org/TR/tabular-metadata/#dfn-table-group-description}{table group description}
}
\description{
The spec defines a \href{https://w3c.github.io/csvw/metadata/#normalization}{normalisation process}.
So far this function just coerces the metadata to ensure it describes a table group.
}
