% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{normalise_property}
\alias{normalise_property}
\title{Normalise an annotation property}
\usage{
normalise_property(property, base_url)
}
\arguments{
\item{property}{an annotation property (a list)}
}
\value{
a property (list) a
}
\description{
This follows the [normalisation](https://w3c.github.io/csvw/metadata/#normalization)
process set out in the csvw specification.
}
