% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{normalise_url}
\alias{normalise_url}
\title{Normalise URL of link annotations}
\usage{
normalise_url(property, base_url)
}
\arguments{
\item{property}{a link annotation (single list element)}

\item{base_url}{the base_url to use for normalisation}
}
\value{
A link annotation with a normalised URL
}
\description{
Such that it specified absolutely with reference to a base
}
