% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{parse_columns}
\alias{parse_columns}
\title{Parse columns schema}
\usage{
parse_columns(columns)
}
\arguments{
\item{columns}{a list of lists specification of columns}
}
\value{
a data frame with a row per column specification
}
\description{
Parse columns schema
}
\keyword{internal}
