% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{parse_metadata}
\alias{parse_metadata}
\title{Parse metadata}
\usage{
parse_metadata(metadata, location)
}
\arguments{
\item{metadata}{a csvw metadata list}

\item{location}{the location of the metadata}
}
\value{
metadata coerced into a
\href{https://www.w3.org/TR/tabular-metadata/#dfn-table-group-description}{table group description}
}
\description{
Coerces the metadata to ensure it describes a table group.
Retrieves any linked tableSchema.
}
