% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{property_type}
\alias{property_type}
\title{Identify the type of an annotation property}
\usage{
property_type(property)
}
\description{
This can be done from the name of the property. The
[metadata descriptor diagram](https://w3c.github.io/csvw/metadata/properties.svg)
provides a succinct summary of the types of each property and how they fit together
}
\details{
@param property a named list element
 @return a string defining the property type
 @examples
 property_type(list(url="http://example.net"))
 @md
}
