% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{read_csvw}
\alias{read_csvw}
\title{Read CSV on the Web}
\usage{
read_csvw(file, metadata = NULL)
}
\arguments{
\item{file}{a path for a csv table or a json metadata document}

\item{metadata}{optional user metadata}
}
\value{
csvw metdata list, with a `dataframe` property added to each table
}
\description{
If the argument to `file` is a json metadata document, this will be used to find csv files for
each table using the value of `csvw:url`.
}
\details{
If the argument to `file` is a csv file, and no metadata is provided, an attempt is made to
derive metadata.

If the argument to `file` is a csv file, and the metadata is provided, then the given csv will
override the value of `csvw:url`.

The csvw metadata is returned as a list. In each table in the table group, an element named
`dataframe` is added which provides the contents of the csv table parsed into a data frame using
the table schema.
}
\examples{
\dontrun{
read_csvw("metadata.json")
read_csvw("table.csv", "metadata.json")
}
}
