% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{read_csvw_dataframe}
\alias{read_csvw_dataframe}
\title{Read a data frame from the first table in a csvw}
\usage{
read_csvw_dataframe(filename, metadata = NULL)
}
\arguments{
\item{filename}{a path for a csv table or a json metadata document}

\item{metadata}{optional user metadata}
}
\value{
A data frame parsed using the table schema
}
\description{
Wrapper around \code{read_csvw} convenient when you're only interested in the data and there's only one table
}
