% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{read_metadata}
\alias{read_metadata}
\title{Read and parse CSVW Metadata}
\usage{
read_metadata(filename)
}
\arguments{
\item{filename}{a path for a json metadata document}
}
\value{
csvw metadata list
}
\description{
Reads in a json document as a list, transforming columns specifications into a dataframe.
}
