% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{render_cell}
\alias{render_cell}
\title{Serialise cell values for JSON representation}
\usage{
render_cell(cell)
}
\arguments{
\item{cell}{a typed value}
}
\value{
a representation comparable with the JSON representation (typically a string)
}
\description{
Serialise cell values for JSON representation
}
\keyword{internal}
