% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{render_uri_templates}
\alias{render_uri_templates}
\title{Render URI templates}
\usage{
render_uri_templates(templates, ...)
}
\arguments{
\item{templates}{a character vector with URI templates}

\item{...}{bindings for variables to be interpolated into templates}
}
\description{
This doesn't yet implement the whole of RFC 6570, just enough to make the tests pass
}
