% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{resolve_url}
\alias{resolve_url}
\title{Resolve one URL against another}
\usage{
resolve_url(url1, url2)
}
\arguments{
\item{url1}{the base url}

\item{url2}{a relative url}
}
\value{
A single absolute url
}
\description{
Resolve one URL against another
}
\keyword{internal}
