% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{set_uri_base}
\alias{set_uri_base}
\title{Set the base of a URI template}
\usage{
set_uri_base(t, url)
}
\arguments{
\item{t}{a character vector of URI templates}

\item{url}{a filename url being used as a context (string)}
}
\value{
a character vector of templates with base paths/ domains set appropriately
}
\description{
Set the base of a URI template
}
