% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{type_to_datatype}
\alias{type_to_datatype}
\title{Map R types to csvw datatype}
\usage{
type_to_datatype(types)
}
\arguments{
\item{types}{a list of R types}
}
\value{
a list of csvw datatypes
}
\description{
Translate R types to \href{https://www.w3.org/TR/tabular-metadata/#datatypes}{csvw datatypes}.
Acts as an inverse of \code{datatype_to_type} but doesn't provide a 1:1 correspondence.
}
