% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_csvw}
\alias{validate_csvw}
\title{Validate CSVW specification}
\usage{
validate_csvw(csvw)
}
\arguments{
\item{csvw}{a csvw metadata specification (a list)}
}
\value{
a validation report (list)
}
\description{
Follows the \href{https://w3c.github.io/csvw/syntax/#validating-tables}{csvw table validation}
procedure.
}
