% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_referential_integrity}
\alias{validate_referential_integrity}
\title{Validate the referential integrity of a csvw table group}
\usage{
validate_referential_integrity(csvw)
}
\arguments{
\item{csvw}{the metadata annotation}
}
\value{
a list specifying any foreign key violations
}
\description{
Fails if foreign keys aren't found in the referenced tables
}
