% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vec_depth}
\alias{vec_depth}
\title{Calculate depth of vector safely}
\usage{
vec_depth(x)
}
\arguments{
\item{x}{a vector}
}
\value{
An integer
}
\description{
Like \code{purrr::vec_depth} but doesn't attempt to descend into errors
}
\keyword{internal}
