% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{add_dataframe}
\alias{add_dataframe}
\title{Add data frame to csvw table annotation}
\usage{
add_dataframe(table, filename, group)
}
\arguments{
\item{table}{a \code{csvw:Table} annotation}

\item{filename}{a filename/ URL for the csv table}

\item{group}{a list of metadata for the table group to use as a fallback}
}
\value{
a table annotation with a \code{dataframe} attribute added with data frame
holding the contents of the table
}
\description{
Add data frame to csvw table annotation
}
\keyword{internal}
