% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{add_dataframe}
\alias{add_dataframe}
\title{Add data frame to csvw table annotation}
\usage{
add_dataframe(table, filename, dialect, group_schema)
}
\arguments{
\item{table}{a \code{csvw:Table} annotation}

\item{filename}{a filename/ URL for the csv table}

\item{dialect}{a list describing the dialect for reading the csv file}

\item{group_schema}{a fallback tableSchema from the table group}
}
\value{
a table annotation with a \code{dataframe} attribute added with data frame
holding the contents of the table
}
\description{
Add data frame to csvw table annotation
}
