% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{datatype_to_type}
\alias{datatype_to_type}
\title{Map csvw datatypes to R types}
\usage{
datatype_to_type(x)
}
\arguments{
\item{x}{a character vector of datatypes}
}
\value{
compact string representation for `readr::cols` with one character per column-type
}
\description{
Map csvw datatypes to R types
}
