% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{is_non_core_annotation}
\alias{is_non_core_annotation}
\title{Determine if an annotation is non-core}
\usage{
is_non_core_annotation(property)
}
\arguments{
\item{property}{a list element}
}
\value{
\code{TRUE} the annotation is core, \code{FALSE} otherwise
}
\description{
Checks if the annotation is non-core, and should thus be treated as a json-ld note.
}
