% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{json_ld_to_json}
\alias{json_ld_to_json}
\title{Convert json-ld annotation to json}
\usage{
json_ld_to_json(property)
}
\arguments{
\item{property}{a json-ld annotation (single list element)}
}
\value{
A compacted list element
}
\description{
Follows the \href{https://w3c.github.io/csvw/csv2json/#json-ld-to-json}{rules for JSON-LD to JSON conversion set out in the csv2json standard}.
}
\keyword{internal}
