% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{locate_table}
\alias{locate_table}
\title{Locate csv data table}
\usage{
locate_table(filename, url)
}
\arguments{
\item{filename}{the file passed to \code{read_csvw} in the first place (could be the csv or json annotations)}

\item{url}{the location of the the table as defined in the metadata}
}
\value{
The location of the table
}
\description{
Locate csv data table
}
\keyword{internal}
