% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{normalise_url}
\alias{normalise_url}
\title{Normalise a URL}
\usage{
normalise_url(url, base)
}
\arguments{
\item{url}{a string}

\item{base}{the base to use for normalisation}
}
\value{
A string containing a normalised URL
}
\description{
Ensures that a url is specified absolutely with reference to a base
}
\keyword{internal}
