% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{property_type}
\alias{property_type}
\title{Identify the type of an annotation property}
\usage{
property_type(property)
}
\arguments{
\item{property}{a named list element}
}
\value{
a string defining the property type
}
\description{
This can be done from the name of the property. The
\href{https://w3c.github.io/csvw/metadata/properties.svg}{metadata descriptor diagram}
provides a succinct summary of the types of each property and how they fit together.
}
\examples{
property_type(list(url="http://example.net"))
}
\keyword{internal}
