% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rmap}
\alias{rmap}
\title{Recursive map}
\usage{
rmap(.x, .f)
}
\arguments{
\item{.x}{a list}

\item{.f}{a function (called with elements of \code{.x} as the first argument)}
}
\value{
A list
}
\description{
Applies function \code{.f} to each element in \code{.x} as per \code{purrr::map}.
If the value of the element is itself a list, then the function is applied to that in turn.
The process is followed recursively until an atomic value at the leaf nodes of the list is found.
}
