% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvwr.R
\name{table_to_list}
\alias{table_to_list}
\title{Convert a table to a list}
\usage{
table_to_list(table, group)
}
\arguments{
\item{table}{the csvw table}

\item{group}{list of metadata for the group used for a fallback schema and dialect}
}
\value{
a list representation of the table's contents
}
\description{
Follows the pattern for csv2json
}
\keyword{internal}
